requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
google-generativeai==0.8.5  # try this; if pip fails, check package name
